+++ SKSSL.Tests/Styles.cs
@@ -14,6 +14,8 @@ public class Styles
     public void TEST_SHEET_LOAD()
     {
         LoadStyles();
+        // WARN: LoadStyles doesn't yield a proper stylesheet return. File likely got deleted or isn't there due to
+        //  the filepath changes. (../)
         Assert.IsNotEmpty(UIColorSheet);
         Assert.IsTrue(UIColorSheet[0].Key.Equals("default"));
     }
